Add files via upload
We will look into the Heart disease data set, and focus mainly on Classification. What are the key features that give us an insight into a patient's likelihood of having heart disease? This data set is small, so the methods that will be used may tend to overfit and or produce inaccurate results. However, this is good practice to explore small data sets as they are more typical in real-world scenarios.