Add files via upload
We will look at predicting whether or not an individual from the titanic wreck dataset survived or not. Using the machine learning algorithm, Decision Tree Classifier as our model. We will avoid overfitting by comparing our test accuracy mean with cross-validation, which is a great way to avoid overfitting. The titanic data tells us a lot about the individual's allocated on the Ship. Let's get to it.